+++ apps/backend/src/modules/notifications/notification.repository.ts
@@ -3,6 +3,7 @@ import { type Repository, type ValueOf } from "~/libs/types/types.js";
 import { type UserModel } from "~/modules/users/users.js";
 
 import { NotificationStatus, NotificationType } from "./libs/enums/enums.js";
+import { type filterQueryParameterToNotificationType } from "./libs/maps/maps.js";
 import { NotificationEntity } from "./notification.entity.js";
 import { type NotificationModel } from "./notification.model.js";
 
@@ -21,7 +22,9 @@ class NotificationRepository implements Repository<NotificationEntity> {
 			ValueOf<typeof NotificationType>,
 			string
 		> = {
+			[NotificationType.NEW_COMMENT]: `${user.userDetails.firstName} ${user.userDetails.lastName} left a comment.`,
 			[NotificationType.NEW_FOLLOWER]: `${user.userDetails.firstName} ${user.userDetails.lastName} started following you.`,
+			[NotificationType.NEW_LIKE]: `${user.userDetails.firstName} ${user.userDetails.lastName} liked your activity.`,
 		};
 
 		return notificationTypeToMessage[type];
@@ -136,11 +139,17 @@ class NotificationRepository implements Repository<NotificationEntity> {
 	public async findAllByReceiverUserId(
 		receiverUserId: number,
 		search: string,
+		type: ValueOf<typeof filterQueryParameterToNotificationType>,
 	): Promise<NotificationEntity[]> {
 		const userNotifications = await this.notificationModel
 			.query()
 			.where({ receiverUserId })
-			.where((builder) => {
+			.andWhere((builder) => {
+				if (type) {
+					void builder.andWhere({ type });
+				}
+			})
+			.andWhere((builder) => {
 				void builder
 					.whereILike("user:userDetails.firstName", `%${search}%`)
 					.orWhereILike("user:userDetails.lastName", `%${search}%`)