+++ apps/frontend/src/libs/enums/enums.ts
@@ -9,6 +9,7 @@ export {
 	EmptyLength,
 	ExceptionMessage,
 	FormatDateType,
+	NotificationFilter,
 	PaginationValue,
 	ServerErrorType,
 } from "@trackmates/shared";